Lo ultimo

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: descamil <descamil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/21 20:51:06 by sergio            #+#    #+#             */
-/*   Updated: 2024/06/15 08:27:44 by descamil         ###   ########.fr       */
+/*   Updated: 2024/11/27 17:04:13 by descamil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
